# opt: implement zero-allocated []byte pool for reusing bytes