---
name: Pull request
about: Propose changes to the code
title: 'zero-allocated []byte pool'
labels: 
assignees: 
---

## 1. Are you opening this pull request for bug-fixes, optimizations or new feature?

Optimizations

## 2. Please describe how these code changes achieve your intention.

**old**

```go
go test -bench=. -run=none -benchtime=3s -benchmem -memprofile=old.mem.pprof
goos: linux
goarch: amd64
pkg: github.com/panjf2000/gnet/pkg/pool/byteslice
cpu: Intel(R) Xeon(R) CPU E3-1230 V2 @ 3.30GHz
BenchmarkByteSlice/Run.N-8              51881646                63.92 ns/op           24 B/op          1 allocs/op
BenchmarkByteSlice/Run.Parallel-8       100000000               34.31 ns/op           24 B/op          1 allocs/op
PASS
ok      github.com/panjf2000/gnet/pkg/pool/byteslice    6.875s
```

![old](https://user-images.githubusercontent.com/4979407/145504683-7e95ce77-afa8-46a3-aa95-3d356609690a.png)

------

**new**

```go
go test -bench=. -run=none -benchtime=3s -benchmem -memprofile=new.mem.pprof
goos: linux
goarch: amd64
pkg: github.com/panjf2000/gnet/pkg/pool/byteslice
cpu: Intel(R) Xeon(R) CPU E3-1230 V2 @ 3.30GHz
BenchmarkByteSlice/Run.N-8              144629972               24.83 ns/op            0 B/op          0 allocs/op
BenchmarkByteSlice/Run.Parallel-8       577906443                5.676 ns/op           0 B/op          0 allocs/op
PASS
ok      github.com/panjf2000/gnet/pkg/pool/byteslice    10.049s
```

![new](https://user-images.githubusercontent.com/4979407/145504711-b95d48f5-fae9-4a07-8737-9734ac44d863.png)

------

[byteslice.mem.pprof.21121010.zip](https://github.com/panjf2000/gnet/files/7689590/byteslice.mem.pprof.21121010.zip)

## 4. Checklist

- [x] I have squashed all insignificant commits.
- [x] I have commented my code for explaining package types, values, functions, and non-obvious lines.
- [x] I have written unit tests and verified that all tests passes (if needed).
- [ ] I have documented feature info on the README (only when this PR is adding a new feature).
- [ ] (optional) I am willing to help maintain this change if there are issues with it later.