@panjf2000 2021-12-10
这个 benchmark 我不太理解，Put() 方法是把 *[]byte 指针放回 sync.Pool，为什么还会每次都分配一个新的 reflect.SliceHeader{}？是不是这个测试有问题？