@fufuok 2021-12-10
可以为现在 dev 分支的 byteslice.go  写一个 benchmark 跑一跑看看.

```go
func BenchmarkByteSlice(b *testing.B) {
	b.ReportAllocs()
	for i := 0; i < b.N; i++ {
		bs := Get(1024)
		Put(bs)
	}
}
```

```shell
go test -bench=. -run=none -benchtime=3s -benchmem -memprofile=mem.pprof
```