@fufuok 2021-12-10
我的理解是要把切片对应底层数组的指针放到池中, 取出来时再用切片来引用, 才能达到复用底层数组的复用目的.
如果要把切片直接放入池, 可以类似这样:

```go
func (p *Pool) Get(size int) []byte {
	// ...
	buf := p.pools[idx].Get().(*[]byte)
	*buf = (*buf)[:size]
	return buf
}

func (p *Pool) Put(buf *[]byte) {
	// ...
	*buf = (*buf)[:cap(*buf)]
	p.pools[idx].Put(buf)
}
```