@panjf2000 2021-12-14
> 我的理解是要把切片对应底层数组的指针放到池中, 取出来时再用切片来引用, 才能达到复用底层数组的复用目的. 如果要把切片直接放入池, 可以类似这样:
> 
> ```go
> func (p *Pool) Get(size int) []byte {
> 	// ...
> 	buf := p.pools[idx].Get().(*[]byte)
> 	*buf = (*buf)[:size]
> 	return buf
> }
> 
> func (p *Pool) Put(buf *[]byte) {
> 	// ...
> 	*buf = (*buf)[:cap(*buf)]
> 	p.pools[idx].Put(buf)
> }
> ```
> 
> 使用时需要用:
> 
> ```go
> data := byteslice.Get(1024)
> copy(*data, "123")
> ```

现在的实现就是因为放入 pool，导致 Put 的入参分配在了堆上了，你修改过后就变成了只把底层数组地址放入 pool，所以 Put 函数的入参 buf 就分配在了栈上，因此 benchmark 显示是 0，但是你这种实现看起来有点 tricky，有在线上大规模验证过吗?