@panjf2000 2021-12-24
对了，忘了问你有执行 `go vet` 检查这部分代码吗？
@fufuok