@fufuok 2021-12-24
`go vet` 执行过检查了.